Combine ConverterCodeGenerator into GenerateConvertersTask for better code organization
- Merge all code generation logic into a single GenerateConvertersTask class
- Support both SBT task mode (no args) and command-line mode (2 args)
- Remove duplicate concern separation between the two classes
- Maintain all existing functionality while reducing code complexity
- SBT task mode works as before for nested.proto testing
- Command-line mode allows flexible descriptor file and output directory specification
